Squashed 'libltfat/' changes from a49b5ba8..29dbe827
29dbe827 Cleanup
924c720c Removed libsndfile dependency in the wavhandler
e9015b7b Fixed forgotten function header
481b6020 Unified permissions
0f2800a1 Muted VS warnings
9f10de97 C++ compatibility changes
0c3a8e25 Updated main docpage
3795be91 Updated CMakeLists

git-subtree-dir: libltfat
git-subtree-split: 29dbe82766dc0e80377c7e6b6eb890b319768a8c

+++ modules/libltfat/src/memalloc.c
@@ -27,7 +27,7 @@ ltfat_set_memory_handler (ltfat_memory_handler_t new_handler)
 #define ALIGNBOUNDARY 64
 
 static void*
-ltfat_aligned_malloc(int size)
+ltfat_aligned_malloc(size_t size)
 {
     void* mem = malloc(size + sizeof(void*) + ALIGNBOUNDARY - 1);
     void* ptr = (void**)
